requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
daphne
psycopg2-binary
whitenoise
requests
python-dotenv 
Pillow
pytest
pytest-django
pytest-asyncio
pytest-cov
Faker
django-storages[s3]
boto3